trc1010: make test work in both x86_64 and aarch64
The old cdk binary is not compatible with aarch64, so we need to
download the correct one for the architecture we are running on.

cdk_linux_amd64 (removed) is v0.1.10; its replacement is v1.0.5, which
is not the latest version, but it is the one that works with the
current test. For more information, see:

https://github.com/cdk-team/CDK/blob/b0ca845156bd9ef8c2d2ce13ab33699f04b9047d/pkg/exploit/mount_cgroup.go#L198-L210